Moved RewriteSourceFileWrapper
